Update Transfer-class (still WIP)
Added the two functions verifyFile() and validateCert(). In save() we
now first make sure that the certificate is existing and not expired.
After downloading the update it gets verified. Potentially bad files get
dropped.

Fetching the signature file is not yet implemented. Further there are
some messages needed for the user in case of an expired/missing
certificate or failed validation.