Reduce HTTP GET requests (Ilch-updater)
Picked the low hanging fruits to reduce the HTTP GET requests to the
update server.
Before these changes a check for updates generated six GET requests. Now
this is down to two, which is a reduction of 2/3 or 66,6 %.

For the case of checking for an update and downloading one we still see
the 2/3 reduction for requests related to checking for updates. Adding
up these requests and two requests for downloading the update and
signature we previously had 14 request - now six, which is a reduction
of around 57 %.

Before:
check for updates + download of update = 12 Requests + 2 for the
download

After:
check for updates + download of update = 4 Requests + 2 for the download

Further reducing the requests needed for an update-check might be worth
the work.