Added deleteAuthToken() and deleteAllAuthTokenOfUser()
Added deleteAuthToken() and deleteAllAuthTokenOfUser(). The first one is
planned to be used when the user logs out. The second one is for the
case when the selector is fine, but the token was wrong. This might be a
theft and therefore delete all authTokens of the affected user.

Check if the authToken is expired and don't use it if that is the case.

Use "fetchAssoc" in getAuthToken() as it is more clean to access the
array later with e.g. "token", "userid" and not with an index.