Translate navigation in userpanel
Please review in detail.

Use $usermenu->getid() and the new function getTransKey() to decide
which translation needs to be displayed.

Removed the title column from [prefix]_user_menu as it looks no longer
used/needed.

Removed the getter/setters for title and text from the usermenu-model.

An alternative implementation would be to use the title to store the
"translationKey".