Remove "user.css" from "navi.php"
The "navi.php" mainly included in application\modules\user\views\panel\
with the exception of
application\modules\newsletter\views\index\settings.php added the
"user.css". In the case of the newsletter-settings this didn't worked
and resulted in a 404-error.

Further the "user.css" is already added in
application\modules\user\views\panel\ and
application\modules\newsletter\views\index\settings.php. Therefore it
was added twice.

By removing the addition of the "user.css" in "navi.php" the 404-error
and duplicated addition of it is fixed.