Added calls to setRowsPerPage() and some adjustements
Call setRowsPerPage() of the pagination-class on every needed location
to use the setting done for the specific module.

Added more settings to the war-module and renamed the setting for the
media-module.

The media-module might need more work as there is a specific value set
for rowsPerPage in iframe.php on line 36, 71 and 104. There need to be a
decision on what to do with these cases.