Add getFormattedCurrency()
Added getFormattedCurrency() to the Translator-class and to the
Base-class. Use this function in the checkout-module.

Adjusted the dafault currency-values added to the database when
installing the checkout-module. Now these are just the currency-codes.

If getFormattedCurrency() fails in applying the provided currency-code
it falls back to just format the number and append the supplied
currency-code. This is for example the case when the currency is
"Bitcoin".