Add curlClose() and a destructor to the transfer-class
Add curlClose() and a destructor to the transfer-class.
Call curlClose() after installing/updating.

The destructor calls curlClose(). This is currenctly the case with the
version info on the admincenter-startpage and an check for ilch-updates
with none found.

"The destructor method will be called as soon as there are no other
references to a particular object, or in any order during the shutdown
sequence."