Add deleting of subscribers
Subscribers/Receivers can now be deleted in the admincenter.

Added deleteSubscriberBySelector() to avoid adding an email address as
parameter to the deleteAction-url. Instead of an email-address the
selector is now added as parameter.

Email addresses or other potential sensible data should not be added as
parameters in urls to avoid them being leaked in browser histories or to
proxies.