Remove unnecessary js language attributes
„W3C did not adopt the language attribute, favoring instead a type
attribute which takes a MIME type. Unfortunately, the MIME type was not
standardized, so it is sometimes "text/javascript" or
"application/ecmascript" or something else. Fortunately, all browsers
will always choose JavaScript as the default programming language, so
it is always best to simply write <script>. It is smallest, and it
works on the most browsers.“