Fix orphaned entries in gallery_imgs when deleting a gallery.
Remove useless column gallery_id in gallery_items table.

Rename cat column to gallery_id in the gallery_imgs table. The id in this column is always one of a gallery and never one of a category.

Add constraint to delete images when the gallery gets deleted.

Don't delete everything, but only not existing items when changing the gallery in the admincenter.